backend/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = ""
authors = [{name = "Pumidol Leelerdsakulvong", email = "pumidol.l@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.0,<6.0",
    "django-ninja>=1.4.0,<2.0",
    "whitenoise>=6.9.0,<7.0",
    "psycopg[binary]>=3.2.0,<4.0",
    "pydantic[email]>=2.11.0,<3.0",
    "django-cors-headers>=4.7.0,<5.0",
    "hypercorn>=0.17.0,<1.0",
    "rich>=14.1.0,<15.0",
    "celery>=5.5.0,<6.0",
    "django-redis>=6.0.0,<7.0",
    "pillow>=11.3.0,<12.0",
    "django-storages[s3]>=1.14.0,<2.0",
    "httpx>=0.28.0,<1.0",
    "discord-webhook>=1.4.0,<2.0",
    "pandas>=2.3.0,<3.0",
    "django-celery-beat>=2.8.0,<3.0",
    "gspread>=6.2.0,<7.0",
    "gspread-dataframe>=4.0.0,<5.0",
]

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
